tests UserController done, application.properties in backend/test/resources added

+++ backend/src/main/java/de/neuefische/teamprojectrestaurantapp/model/Order.java
@@ -23,6 +23,6 @@ public class Order {
 
     private String orderNumber;
     private String orderInformation; // Informationen kommen aus dem Frontend
-    private int totalPrice;
+    private float totalPrice;
 
 }
